split routes_api
